+++ src/platform.matter.test.ts
@@ -492,6 +492,7 @@ describe('Matterbridge ' + NAME, () => {
     expect(aggregator.parts.has(device)).toBeTruthy();
     expect(aggregator.parts.has(device.id)).toBeTruthy();
     expect(subscribeAttributeSpy).toHaveBeenCalledTimes(0);
+    expect(child.getAttribute(BooleanState.Cluster.id, 'stateValue')).toBe(false); // Contact Sensor: true = closed or contact, false = open or no contact
 
     jest.clearAllMocks();
     await haPlatform.onConfigure();
@@ -513,6 +514,80 @@ describe('Matterbridge ' + NAME, () => {
     haPlatform.ha.hassDevices.delete(contactDevice.id);
     haPlatform.ha.hassEntities.delete(contactDeviceEntity.entity_id);
     haPlatform.ha.hassStates.delete(contactDeviceEntityState.entity_id);
+    await new Promise((resolve) => setTimeout(resolve, 100)); // Wait for async storage number persist operations to complete
+    await device.delete();
+    expect(aggregator.parts.size).toBe(0);
+  });
+
+  it('should call onStart and register a Leak device', async () => {
+    const leakDevice = {
+      area_id: null,
+      disabled_by: null,
+      entry_type: null,
+      id: 'd80898f83188759ed7329e97df00ee7c',
+      labels: [],
+      name: 'Leak Sensor',
+      name_by_user: null,
+    } as unknown as HassDevice;
+
+    const leakDeviceEntity = {
+      area_id: null,
+      device_id: leakDevice.id,
+      entity_category: null,
+      entity_id: 'binary_sensor.leak_sensor',
+      has_entity_name: true,
+      id: '0b25a337cb83edefb1d310450ad2b0ac',
+      labels: [],
+      name: null,
+      original_name: 'Leak Sensor',
+    } as unknown as HassEntity;
+
+    const leakDeviceEntityState = {
+      entity_id: leakDeviceEntity.entity_id,
+      state: 'off', // 'on' for leak, 'off' for no leak
+      attributes: { device_class: 'moisture', friendly_name: 'Leak Sensor' },
+    } as unknown as HassState;
+
+    haPlatform.ha.hassDevices.set(leakDevice.id, leakDevice);
+    haPlatform.ha.hassEntities.set(leakDeviceEntity.entity_id, leakDeviceEntity);
+    haPlatform.ha.hassStates.set(leakDeviceEntityState.entity_id, leakDeviceEntityState);
+
+    await haPlatform.onStart('Test reason');
+    await new Promise((resolve) => setTimeout(resolve, 100)); // Wait for async operations to complete
+    expect(mockLog.info).toHaveBeenCalledWith(`Starting platform ${idn}${mockConfig.name}${rs}${nf}: Test reason`);
+    expect(mockMatterbridge.addBridgedEndpoint).toHaveBeenCalledTimes(1);
+    expect(haPlatform.matterbridgeDevices.size).toBe(1);
+    expect(haPlatform.matterbridgeDevices.get(leakDevice.id)).toBeDefined();
+    device = haPlatform.matterbridgeDevices.get(leakDevice.id) as MatterbridgeEndpoint;
+    expect(device.construction.status).toBe(Lifecycle.Status.Active);
+    const child = device?.getChildEndpointByName(leakDeviceEntity.entity_id.replace('.', ''));
+    expect(child).toBeDefined();
+    if (!child) return;
+    await child.construction.ready;
+    expect(child.construction.status).toBe(Lifecycle.Status.Active);
+    expect(aggregator.parts.has(device)).toBeTruthy();
+    expect(aggregator.parts.has(device.id)).toBeTruthy();
+    expect(subscribeAttributeSpy).toHaveBeenCalledTimes(0);
+    expect(child.getAttribute(BooleanState.Cluster.id, 'stateValue')).toBe(false); // Water Leak Detector: true = leak, false = no leak
+
+    jest.clearAllMocks();
+    await haPlatform.onConfigure();
+    await new Promise((resolve) => setTimeout(resolve, 100)); // Wait for async updateHandler operations to complete
+    expect(mockLog.debug).toHaveBeenCalledWith(expect.stringContaining(`Configuring state ${CYAN}${leakDeviceEntityState.entity_id}${db} for device ${CYAN}${leakDevice.id}${db}`));
+    expect(setAttributeSpy).toHaveBeenCalledWith(BooleanState.Cluster.id, 'stateValue', false, expect.anything()); // Water Leak Detector: true = leak, false = no leak
+
+    jest.clearAllMocks();
+    haPlatform.updateHandler(leakDevice.id, leakDeviceEntityState.entity_id, leakDeviceEntityState, { ...leakDeviceEntityState, state: 'on' }); // 'on' for leak, 'off' for no leak
+    expect(setAttributeSpy).toHaveBeenCalledWith(BooleanState.Cluster.id, 'stateValue', true, expect.anything()); // Water Leak Detector: true = leak, false = no leak
+
+    jest.clearAllMocks();
+    haPlatform.updateHandler(leakDevice.id, leakDeviceEntityState.entity_id, leakDeviceEntityState, { ...leakDeviceEntityState, state: 'off' }); // 'on' for leak, 'off' for no leak
+    expect(setAttributeSpy).toHaveBeenCalledWith(BooleanState.Cluster.id, 'stateValue', false, expect.anything()); // Water Leak Detector: true = leak, false = no leak
+
+    haPlatform.matterbridgeDevices.delete(leakDevice.id);
+    haPlatform.ha.hassDevices.delete(leakDevice.id);
+    haPlatform.ha.hassEntities.delete(leakDeviceEntity.entity_id);
+    haPlatform.ha.hassStates.delete(leakDeviceEntityState.entity_id);
     await new Promise((resolve) => setTimeout(resolve, 100)); // Wait for async storage number persist operations to complete
     await device.delete();
     expect(aggregator.parts.size).toBe(0);