+++ CHANGELOG.md
@@ -23,6 +23,7 @@ If you like this project and find it useful, please consider giving it a star on
 ### Fixed
 
 - [reconnect]: Added missed call to fetchData and subscribe on reconnect.
+- [startup]: Added correct value for BooleanState cluster to avoid controller alarms.
 
 <a href="https://www.buymeacoffee.com/luligugithub">
   <img src="bmc-button.svg" alt="Buy me a coffee" width="80">